Funções /sku/notify e /sku/priceRange

+++ src/main/java/com/epicom/Repository/SkuRepository.java
@@ -8,6 +8,6 @@ import java.util.Optional;
 /**
  * Created by Julio on 18/11/2016.
  */
-public interface SkuRepository extends JpaRepository<Sku, Long> {
+public interface SkuRepository extends JpaRepository<Sku, Integer> {
     Optional<Sku> findByProductId(String productId);
 }
